Feedback-system-project-master/pyproject/requirements.txt
better-profanity==0.7.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
nltk==3.6.7
protobuf==3.19.1
pyenchant==3.2.2
regex==2021.11.10
tqdm==4.62.3
Werkzeug==2.0.2
